requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
